fixed from PVS-Studio
V668 There is no sense in testing the 'lpPropTagsV2' pointer against null, as the memory was allocated using the 'new' operator. The exception will be generated in the case of memory allocation error. addins.cpp 66
V668 There is no sense in testing the 'lpResults->lpDialogControlResults' pointer against null, as the memory was allocated using the 'new' operator. The exception will be generated in the case of memory allocation error. addins.cpp 945
V668 There is no sense in testing the 'lpResults' pointer against null, as the memory was allocated using the 'new' operator. The exception will be generated in the case of memory allocation error. addins.cpp 941