Trying to delete a body from an OFT, get MAPI_E_NOT_FOUND.
This is because the property we have is of type PT_ERROR. We can change that to PT_UNSPECIFIED, which works for some stores, but some (like OFT) still don't like this.

Fix is to watch for MAPI_E_NOT_FOUND from a deletion attempt, then try the common stream types just in case they work.

Fixes #108